+++ Client/Write/WritePrecision.cs
@@ -1,3 +1,5 @@
+using System;
+
 namespace InfluxDB3.Client.Write;
 
 /// <summary>
@@ -24,4 +26,31 @@ public enum WritePrecision
     /// Enum Ns for value: ns
     /// </summary>
     Ns = 4
+}
+
+public static class WritePrecisionConverter
+{
+    public static string ToV2ApiString(WritePrecision precision)
+    {
+        return precision switch
+        {
+            WritePrecision.Ns => "ns",
+            WritePrecision.Us => "us",
+            WritePrecision.Ms => "ms",
+            WritePrecision.S => "s",
+            _ => throw new ArgumentException($"Unsupported precision '{precision}'"),
+        };
+    }
+
+    public static string ToV3ApiString(WritePrecision precision)
+    {
+        return precision switch
+        {
+            WritePrecision.Ns => "nanosecond",
+            WritePrecision.Us => "microsecond",
+            WritePrecision.Ms => "millisecond",
+            WritePrecision.S => "second",
+            _ => throw new ArgumentException($"Unsupported precision '{precision}'"),
+        };
+    }
 }