+++ CHANGELOG.md
@@ -7,6 +7,16 @@
 1. [#155](https://github.com/InfluxCommunity/influxdb3-csharp/pull/155): Allows setting grpc options.
 1. [#157](https://github.com/InfluxCommunity/influxdb3-csharp/pull/157): Fix: always clone `DefaultOptions` to keep it
    immutable.
+1. [#158](https://github.com/InfluxCommunity/influxdb3-csharp/pull/158): Support fast writes without waiting for WAL
+   persistence:
+    - New write option (`WriteOptions.NoSync`) added: `true` value means faster write but without the confirmation that
+      the data was persisted. Default value: `false`.
+    - **Supported by self-managed InfluxDB 3 Core and Enterprise servers only!**
+    - Also configurable via connection string query parameter (`writeNoSync`).
+    - Also configurable via environment variable (`INFLUX_WRITE_NO_SYNC`).
+    - Long precision string values added from v3 HTTP API: `"nanosecond"`, `"microsecond"`, `"millisecond"`,
+      `"second"` (
+      in addition to the existing `"ns"`, `"us"`, `"ms"`, `"s"`).
 
 ## 1.1.0 [2025-03-26]